requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
tenacity>=8.2.3
structlog>=23.2.0
pyyaml>=6.0.1
schedule>=1.2.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
itsdangerous>=2.1.2
# pyngrok>=5.0.0  # Not needed in production - only for development tunnels
cloudinary>=1.36.0
openai>=1.0.0
bcrypt>=4.0.0
cryptography>=42.0.0